utils.py: Move send_email over from sct_utils
