+++ spinalcordtoolbox/utils.py
@@ -256,6 +256,60 @@ def tmp_create(basename=None):
     logger.info("Creating temporary folder (%s)" % tmpdir)
     return tmpdir
 
+def send_email(addr_to, addr_from, passwd, subject, message='', filename=None, html=False, smtp_host=None, smtp_port=None, login=None):
+    if smtp_host is None:
+        smtp_host = os.environ.get("SCT_SMTP_SERVER", "smtp.gmail.com")
+    if smtp_port is None:
+        smtp_port = int(os.environ.get("SCT_SMTP_PORT", 587))
+    if login is None:
+        login = addr_from
+
+    import smtplib
+    from email.mime.multipart import MIMEMultipart
+    from email.mime.text import MIMEText
+    from email.mime.base import MIMEBase
+    from email import encoders
+
+    if html:
+        msg = MIMEMultipart("alternative")
+    else:
+        msg = MIMEMultipart()
+
+    msg['From'] = addr_from
+    msg['To'] = addr_to
+    msg['Subject'] = subject
+
+    body = message
+    if not isinstance(body, bytes):
+        body = body.encode("utf-8")
+
+    body_html = """
+<html><pre style="font: monospace"><body>
+{}
+</body></pre></html>""".format(body).encode()
+
+    if html:
+        msg.attach(MIMEText(body_html, 'html', "utf-8"))
+
+    msg.attach(MIMEText(body, 'plain', "utf-8"))
+
+    # filename = "NAME OF THE FILE WITH ITS EXTENSION"
+    if filename:
+        attachment = open(filename, "rb")
+        part = MIMEBase('application', 'octet-stream')
+        part.set_payload((attachment).read())
+        encoders.encode_base64(part)
+        part.add_header('Content-Disposition', "attachment; filename= %s" % filename)
+        msg.attach(part)
+
+    # send email
+    server = smtplib.SMTP(smtp_host, smtp_port)
+    server.starttls()
+    server.login(login, passwd)
+    text = msg.as_string()
+    server.sendmail(addr_from, addr_to, text)
+    server.quit()
+
 
 def __get_branch():
     """
